Move theming resource tests to test_theme_utils
read_theming_resource lives in theme_utils but was tested from
test_design_tokens, which made the resource loader look like an
implementation detail of design tokens. Collect its tests in a module
matching the one under test, and cover read_design_token_file, which
had no direct tests.

The load_tokens tests now stub read_theming_resource directly instead
of faking importlib.resources, which removes the resource stub classes.
